requirements.txt
flask
pandas
openpyxl
google-generativeai
markdown
gunicorn
